fix up x86 cleanup

+++ source/installer.py
@@ -858,6 +858,7 @@ def install(shouldCreateDesktopShortcut: bool = True, shouldRunAtLogon: bool = T
 	installDir = WritePaths.defaultInstallDir
 	installDirX86 = WritePaths._installDirX86
 	startMenuFolder = WritePaths.defaultStartMenuFolder
+	shouldCleanX86 = installDirX86 is not None
 	# Give some time for the installed NVDA (which may have been running on a secure screen)
 	# to shut down before we start deleting files.
 	time.sleep(1)
@@ -877,17 +878,19 @@ def install(shouldCreateDesktopShortcut: bool = True, shouldRunAtLogon: bool = T
 		numTries=6,
 		retryWaitInterval=0.5,
 	)
-	_deleteFileGroupOrFail(
-		installDirX86,
-		_nvdaExes.union({"nvda_service.exe", "nvda_eoaProxy.exe"}),
-		numTries=6,
-		retryWaitInterval=0.5,
-	)
+	if shouldCleanX86:
+		_deleteFileGroupOrFail(
+			installDirX86,
+			_nvdaExes.union({"nvda_service.exe", "nvda_eoaProxy.exe"}),
+			numTries=6,
+			retryWaitInterval=0.5,
+		)
 	unregisterInstallation(keepDesktopShortcut=shouldCreateDesktopShortcut)
 	if prevInstallPath:
 		removeOldLoggedFiles(prevInstallPath)
 	removeOldProgramFiles(installDir)
-	removeOldProgramFiles(installDirX86)
+	if shouldCleanX86:
+		removeOldProgramFiles(installDirX86)
 	copyProgramFiles(installDir)
 	for f in ("nvda_UIAccess.exe", "nvda_noUIAccess.exe"):
 		f = os.path.join(installDir, f)
@@ -897,7 +900,8 @@ def install(shouldCreateDesktopShortcut: bool = True, shouldRunAtLogon: bool = T
 	else:
 		raise RuntimeError("No available executable to use as nvda.exe")
 	removeOldLibFiles(installDir, rebootOK=True)
-	removeOldLibFiles(installDirX86, rebootOK=True)
+	if shouldCleanX86:
+		removeOldLibFiles(installDirX86, rebootOK=True)
 	registerInstallation(
 		installDir,
 		startMenuFolder,
